+++ index.js
@@ -1,2 +1,2 @@
-!function(t,e){"object"==typeof exports&&"undefined"!=typeof module?module.exports=e(require("mobx-state-tree")):"function"==typeof define&&define.amd?define(["mobx-state-tree"],e):t["trampss-mst-onaction"]=e(t.mobxStateTree)}(this,function(t){"use strict";!function(){function t(t){this.value=t}function e(e){function n(o,u){try{var i=e[o](u),a=i.value;a instanceof t?Promise.resolve(a.value).then(function(t){n("next",t)},function(t){n("throw",t)}):r(i.done?"return":"normal",i.value)}catch(t){r("throw",t)}}function r(t,e){switch(t){case"return":o.resolve({value:e,done:!0});break;case"throw":o.reject(e);break;default:o.resolve({value:e,done:!1})}(o=o.next)?n(o.key,o.arg):u=null}var o,u;this._invoke=function(t,e){return new Promise(function(r,i){var a={key:t,arg:e,resolve:r,reject:i,next:null};u?u=u.next=a:(o=u=a,n(t,e))})},"function"!=typeof e.return&&(this.return=void 0)}"function"==typeof Symbol&&Symbol.asyncIterator&&(e.prototype[Symbol.asyncIterator]=function(){return this}),e.prototype.next=function(t){return this._invoke("next",t)},e.prototype.throw=function(t){return this._invoke("throw",t)},e.prototype.return=function(t){return this._invoke("return",t)}}();var e=Object.assign||function(t){for(var e=1;e<arguments.length;e++){var n=arguments[e];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(t[r]=n[r])}return t},n=function(n){var r=n.name,o=n.type,u=n.context,i=n.args,a=t.getPath(u),c={fullpath:a+"/"+r,path:a,name:r,args:i};switch(o){case"process_return":return e({},c,{ended:!0});case"action":return c;default:return}};return function(t){return function(e,r){var o=n(e);return o&&t(o,e.tree),r(e)}}});
+!function(t,e){"object"==typeof exports&&"undefined"!=typeof module?e(exports,require("lodash"),require("mobx-state-tree")):"function"==typeof define&&define.amd?define(["exports","lodash","mobx-state-tree"],e):e(t["trampss-mst-onaction"]={},t.lodash,t.mobxStateTree)}(this,function(t,e,n){"use strict";var r=function(t,n){return function(r,o){return!!(e.isString(r)&&r===t.fullpath||e.isFunction(r)&&r(t,n)||e.isRegExp(r)&&t.fullpath.match(r))&&o(t,n)}};r.ended=function(t){return function(e,n){return e.ended?t(e,n):function(){return!1}}}(r);!function(){function t(t){this.value=t}function e(e){function n(o,u){try{var i=e[o](u),a=i.value;a instanceof t?Promise.resolve(a.value).then(function(t){n("next",t)},function(t){n("throw",t)}):r(i.done?"return":"normal",i.value)}catch(t){r("throw",t)}}function r(t,e){switch(t){case"return":o.resolve({value:e,done:!0});break;case"throw":o.reject(e);break;default:o.resolve({value:e,done:!1})}(o=o.next)?n(o.key,o.arg):u=null}var o,u;this._invoke=function(t,e){return new Promise(function(r,i){var a={key:t,arg:e,resolve:r,reject:i,next:null};u?u=u.next=a:(o=u=a,n(t,e))})},"function"!=typeof e.return&&(this.return=void 0)}"function"==typeof Symbol&&Symbol.asyncIterator&&(e.prototype[Symbol.asyncIterator]=function(){return this}),e.prototype.next=function(t){return this._invoke("next",t)},e.prototype.throw=function(t){return this._invoke("throw",t)},e.prototype.return=function(t){return this._invoke("return",t)}}();var o=Object.assign||function(t){for(var e=1;e<arguments.length;e++){var n=arguments[e];for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(t[r]=n[r])}return t},u=function(t){var e=t.name,r=t.type,u=t.context,i=t.args,a=n.getPath(u),c={fullpath:a+"/"+e,path:a,name:e,args:i};switch(r){case"process_return":return o({},c,{ended:!0});case"action":return c;default:return}};t.default=function(t){return function(e,n){var r=u(e);return r&&t(r,e.tree),n(e)}},t.take=r,Object.defineProperty(t,"__esModule",{value:!0})});
 //# sourceMappingURL=index.js.map
